Convert react-app to Next project format
