+++ auth/auth0-provider-with-history.js
@@ -1,22 +1,31 @@
-import React from "react";
-import { useHistory } from "react-router-dom";
+import React, { useEffect, useState } from "react";
+import Router from "next/router";
 import { Auth0Provider } from "@auth0/auth0-react";
 
+// TODO: Fix redirectUri instead of {window.location.origin} and 
+// Log In button doesn't navigate to domain, but to Undefined
+
 const Auth0ProviderWithHistory = ({ children }) => {
-  const history = useHistory();
+  
+  useEffect(() => {
+    const windowPath = window.location.pathname
+  }, [])
+
   const domain = process.env.REACT_APP_AUTH0_DOMAIN;
   const clientId = process.env.REACT_APP_AUTH0_CLIENT_ID;
   const audience = process.env.REACT_APP_AUTH0_AUDIENCE;
+  const serverUrl = process.env.REACT_APP_SERVER_URL;
 
+  
   const onRedirectCallback = (appState) => {
-    history.push(appState?.returnTo || window.location.pathname);
+    Router.push(appState?.returnTo || windowPath);
   };
 
   return (
     <Auth0Provider
       domain={domain}
       clientId={clientId}
-      redirectUri={window.location.origin}
+      redirectUri={serverUrl}
       onRedirectCallback={onRedirectCallback}
       audience={audience}
     >
