+++ pages/index.js
@@ -1,23 +1,87 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./app";
-import { BrowserRouter as Router } from "react-router-dom";
-import ThemeModeProvider from './context/ThemeModeProvider';
-import Auth0ProviderWithHistory from "./auth/auth0-provider-with-history";
-import history from "./history"
+import React, { useEffect, useState } from "react";
+import {
+  Grid,
+  Paper,
+  Box,
+  Typography,
+  ListItem,
+  List,
+} from "@material-ui/core";
+import JSONPretty from "react-json-pretty";
+
+import Welcome from "../components/welcome";
+import useAxios from "../hooks/useAxios";
+import { backendApi } from "../apis/axiosRequests";
+import ProductsButton from "../components/ProductsButton";
+
+// TODO: Fix JSON Data Output, Possibly change REST API url to be
+// /api/v1/products so that way the frontend and backend don't get confused
+// when making API calls.
 
 
-if (module.hot) {
-  module.hot.accept();
-}
+const Home = () => {
+  const [url, setUrl] = useState("/products");
 
-ReactDOM.render(
-  <Router history={history}>
-    <ThemeModeProvider>
-      <Auth0ProviderWithHistory>
-        <App />
-      </Auth0ProviderWithHistory>
-    </ThemeModeProvider>
-  </Router>,
-  document.querySelector("#root")
-);
+  const { response, isLoading } = useAxios({
+    api: backendApi,
+    method: "get",
+    url: `${url}`,
+    // config: JSON.stringify({ requireAuthentication: true }),
+  });
+
+  const [data, setData] = useState([]);
+
+  useEffect(() => {
+    if (response !== null) {
+      setData(response);
+    }
+  }, [response]);
+
+  return (
+    <Box mt={4}>
+      <Welcome />
+      <Grid container spacing={1}>
+        <Grid item>
+          <ProductsButton
+            text="All Products"
+            onClick={() => setUrl("/products")}
+          />
+        </Grid>
+        <Grid item>
+          <ProductsButton
+            text="Mens Apparel"
+            onClick={() => setUrl("/collections/mens-apparel")}
+          />
+        </Grid>
+        <Grid item>
+          <ProductsButton
+            text="Womens Apparel"
+            onClick={() => setUrl("/collections/womens-apparel")}
+          />
+        </Grid>
+        <Grid item>
+          <ProductsButton
+            text="Holiday"
+            onClick={() => setUrl("/collections/holiday")}
+          />
+        </Grid>
+        <Grid item>
+          <ProductsButton
+            text="Misc"
+            onClick={() => setUrl("/collections/misc")}
+          />
+        </Grid>
+      </Grid>
+      <Box mt={2}>
+        <Typography variant="h5">Example JSON Data Output:</Typography>
+      </Box>
+      <Paper elevation={3}>
+        <Box padding={2}>
+          <JSONPretty id="json-pretty" data={data} style={{ overflow: 'auto' }}></JSONPretty>
+        </Box>
+      </Paper>
+    </Box>
+  );
+};
+
+export default Home;
